==> 01-Data_Prep/01-Census_Data.ipynb <==
## Census Data Retrieval
---
In this Jupyter notebook, we retrieve census data on every zip code and city in the United States.

### Dependencies

# Dependencies
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import requests
import time 
from census import Census
from us import states
import sqlite3
from sqlalchemy import create_engine
import pymysql
pymysql.install_as_MySQLdb()

# Census API Key
c = Census("85ac64b6b5a9c0901b00329d1ef41f0c53ccfc98", year=2015)

### Retrieve Census Data

# Retrieve Census Data (Reference: https://gist.github.com/afhaque/60558290d6efd892351c4b64e5c01e9b)
zip_census = c.acs5.get(( "B19013_001E",
                          "B19301_001E",
                          "B15003_002E",
                          "B15003_017E",
                          "B15003_018E",
                          "B15003_022E",
                          "B15003_021E",
                          "B15003_023E",
                          "B15003_024E",
                          "B15003_025E",
                          "B17001_002E",
                          "B23025_002E",
                          "B23025_005E",
                          "C24010_007E",
                          "C24010_043E",
                          "B01002_001E",
                          "B01002_002E",
                          "B01002_003E",
                          "B01003_001E",
                          "B25064_001E",
                          "B25077_001E",
                          "B25077_001E",
                          "B02001_002E",
                          "B02001_003E",
                          "B02001_004E",
                          "B02001_005E",
                          "B03001_003E"), {'for': 'zip code tabulation area:*'})

# Convert to DataFrame
zip_census = pd.DataFrame(zip_census)

# Column Reordering
zip_census = zip_census.rename(columns={"B19013_001E": "Household Income", 
                                      "B19301_001E": "Income Per Capita",
                                      "B15003_002E": "Education None",
                                      "B15003_017E": "Education High School",
                                      "B15003_018E": "Education GED",
                                      "B15003_022E": "Education Bachelors",
                                      "B15003_021E": "Education Associates",
                                      "B15003_023E": "Education Masters",
                                      "B15003_024E": "Education Professional",
                                      "B15003_025E": "Education Doctorate",
                                      "B17001_002E": "Poverty",
                                      "B23025_002E": "Employment Labor Force",
                                      "B23025_005E": "Employment Unemployed",
                                      "C24010_007E": "Employment Male Computer Engineering",
                                      "C24010_043E": "Employment Female Computer Engineering",
                                      "B01002_001E": "Median Age",
                                      "B01002_002E": "Median Male Age",
                                      "B01002_003E": "Median Female Age",
                                      "B01003_001E": "Population",
                                      "B25064_001E": "Median Gross Rent",
                                      "B25077_001E": "Median Home Value",
                                      "B02001_002E": "White Population",
                                      "B02001_003E": "Black Population",
                                      "B02001_004E": "Native American Population",
                                      "B02001_005E": "Asian Population",
                                      "B03001_003E": "Hispanic Population",
                                      "zip code tabulation area": "zipcode"})

# Visualize Data
zip_census.head()

zip_census.columns

zip_census.count()

### Retrieve Geocoding Data

zip_lat_lng = pd.read_excel("Raw/Latitude_Longitude_Zip.xlsx", converters={"zipcode": str})
zip_lat_lng.head()

zip_lat_lng.count()

### Merge Geocode Data with Census

zip_census_geocoded = zip_census.merge(zip_lat_lng, on="zipcode", how="left")
zip_census_geocoded.head()

zip_census_geocoded.count()

### Add City, State Key

zip_census_geocoded["CityState"] = zip_census_geocoded["city"] + ", " + zip_census_geocoded["state"]
zip_census_geocoded["CityState"].head()

zip_census_geocoded[zip_census_geocoded["CityState"] == "MIAMI, FL"][["Population", "zipcode"]].sum()

### Export CSV of Census Data (By Zip)

zip_census_geocoded.to_csv("Outputs/01-Zip_Census.csv", index=False)

### Convert Columns to Numerics

zip_census_geocoded[['Median Age', 'Median Male Age', 'Median Female Age', 'Population',
       'White Population', 'Black Population', 'Native American Population',
       'Asian Population', 'Hispanic Population', 'Education None',
       'Education High School', 'Education GED', 'Education Associates',
       'Education Bachelors', 'Education Masters', 'Education Professional',
       'Education Doctorate', 'Poverty', 'Household Income',
       'Income Per Capita', 'Employment Labor Force', 'Employment Unemployed',
       'Median Gross Rent', 'Median Home Value',
       'Employment Male Computer Engineering',
       'Employment Female Computer Engineering', 'lat', 'lng']] = zip_census_geocoded[['Median Age', 'Median Male Age', 'Median Female Age', 'Population',
       'White Population', 'Black Population', 'Native American Population',
       'Asian Population', 'Hispanic Population', 'Education None',
       'Education High School', 'Education GED', 'Education Associates',
       'Education Bachelors', 'Education Masters', 'Education Professional',
       'Education Doctorate', 'Poverty', 'Household Income',
       'Income Per Capita', 'Employment Labor Force', 'Employment Unemployed',
       'Median Gross Rent', 'Median Home Value',
       'Employment Male Computer Engineering',
       'Employment Female Computer Engineering', 'lat', 'lng']].apply(pd.to_numeric)

# Check Data Types
zip_census_geocoded.dtypes

### Aggregate at City Level

# Group by city and calculate the sums 
zip_city_sums = zip_census_geocoded.groupby(["CityState", "city", "state"])[["Population", 
                                                                "White Population", 
                                                                "Black Population",
                                                                "Native American Population",
                                                                "Asian Population",
                                                                "Hispanic Population",
                                                                "Education None",
                                                                "Education High School",
                                                                "Education GED",
                                                               "Education Associates",
                                                               "Education Bachelors",
                                                               "Education Masters",
                                                               "Education Professional",
                                                               "Education Doctorate",
                                                               "Poverty",
                                                               "Employment Labor Force",
                                                               "Employment Unemployed",
                                                               "Employment Male Computer Engineering",
                                                               "Employment Female Computer Engineering"]].sum()

# Reset the index
zip_city_sums = zip_city_sums.reset_index()

# Group by city and calculate the averages 
zip_city_means = zip_census_geocoded.groupby(["city", "state"])[["Median Age",
                                                                 "Median Male Age",
                                                                 "Median Female Age",
                                                                 "Household Income",
                                                                 "Income Per Capita",
                                                                 "Median Gross Rent",
                                                                 "Median Home Value", "lat", "lng"
                                                                ]].mean()

# Reset the index
zip_city_means = zip_city_means.reset_index()

zip_city_sums.columns

zip_city_means.columns

city_census_data = zip_city_sums.merge(zip_city_means, on=["city", "state"], how="left")

city_census_data.columns

### Compare the Zip Stats and the City Stats

print(zip_census_geocoded["Population"].sum())
print(zip_census_geocoded["Employment Male Computer Engineering"].sum())
print(zip_census_geocoded["Household Income"].mean())
print(zip_census_geocoded["Median Age"].mean())

print(city_census_data["Population"].sum())
print(city_census_data["Employment Male Computer Engineering"].sum())
print(city_census_data["Household Income"].mean())
print(city_census_data["Median Age"].mean())

### Export City Level CSV

city_census_data.to_csv("Outputs/02-City_Census.csv", index=False)

### Export City and Zip Data to SQLite

conn = sqlite3.connect("Opportunity_Map.db")

city_census_data.to_sql("City_Census", conn, if_exists="replace", index=False)

zip_census_geocoded.to_sql("Zip_Census", conn, if_exists="replace", index=False)

pd.read_sql("select * from Zip_Census", conn)

# Check Tables in SQLite
cur = conn.cursor() 
res = cur.execute("SELECT name FROM sqlite_master WHERE type='table' ORDER BY name;")
for name in res:
    print(name[0])

### Export to Cloud SQL

engine = create_engine('mysql+mysqldb://trilogy:test@35.227.28.228/mapping_data?unix_socket=/cloudsql/sql-projects:us-east1:opportunity-db')
conn = engine.connect()

city_census_data.to_sql("City_Census", conn, if_exists="replace", index=False)
zip_census_geocoded.to_sql("Zip_Instances", conn, if_exists="replace", index=False)
