==> 01-Data_Prep/02-Implement_Rules.ipynb <==
## Implement Rules
---
In this Jupyter notebook, we implement additive rules on the state and university levels, and negative rules on the university level.

### Dependencies

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import requests
import time 
import sqlite3
import csv
from geopy.distance import vincenty
import glob
from sqlalchemy import create_engine
import pymysql
pymysql.install_as_MySQLdb()

### Read the Additive Rules Files

#import border rules csv, clean up
border_rules = {}
csvpath = 'Raw/border-rules.csv'

with open(csvpath, newline= "") as csvfile:
    csvreader = csv.reader(csvfile, delimiter=',')
    next(csvreader)

    for row in csvreader:
        state = row[0]
        border_rules[state] = row[1:]
        
for item in border_rules:
    border_rules[item] = [x for x in border_rules[item] if x]

#import university rules csv, clean up
uni_rules = {}
csvpath2 = 'Raw/uni-rules.csv'

with open(csvpath2, newline= "") as csvfile2:
    csvreader2 = csv.reader(csvfile2, delimiter=',')
    next(csvreader2)

    for row in csvreader2:
        uni = row[0]
        uni_rules[uni] = row[1:]

for item in uni_rules:
    uni_rules[item] = [x for x in uni_rules[item] if x]

border_rules

uni_rules

### Read the List of Physical Locations

#import physical location data for each university to check (make sure uni_rules.csv is updated according to this list)
physical_locations = pd.read_excel("Raw/Physical_Campuses_geo.xlsx")
physical_locations.count()

physical_locations

### Read the Full List of Cities

#import census data for each CityState
conn = sqlite3.connect("Opportunity_Map.db")
full_cities = pd.read_sql("select * from City_Census", conn)

### Run Additive Rules

#create dataframe by checking universities using border rules
border_rules_check = pd.DataFrame(columns=['City', 'State', 'City Lat', 'City Lng',
                                      'University', 'Uni Lat', 'Uni Lng', 'Distance'])

counter = 0
for index, row in physical_locations.iterrows():
     print("Checking " + row["University"])
     state_pull = row["State"]
     uni_lat = row["Uni_lat"]
     uni_lng = row["Uni_lng"]
     uni_name = row["University"]
     uni_loc = row["CityState"]
     for item in border_rules[state_pull]:
         if item is not "":
             #check against the city/state dataframe, create new dataframe of values, add values for city/state and lat/lng
             for index, row in full_cities.iterrows():
                 if row["state"] == item:
                     city_name = row["city"]
                     state_name = row["state"]
                     city_lat = row["lat"]
                     city_lng = row["lng"]
                     city_state = row["CityState"]
                    
                     p1 = (uni_lat, uni_lng)
                     p2 = (city_lat, city_lng)
                     mi_distance = vincenty(p1, p2).miles
                     
                     border_rules_check.set_value(counter, "Distance", mi_distance)
                     border_rules_check.set_value(counter, "City", city_name)
                     border_rules_check.set_value(counter, "State", state_name)
                     border_rules_check.set_value(counter, "City Lat", city_lat)
                     border_rules_check.set_value(counter, "City Lng", city_lng)
                     border_rules_check.set_value(counter, "University", uni_name)
                     border_rules_check.set_value(counter, "Uni CityState", uni_loc)
                     border_rules_check.set_value(counter, "Uni Lat", uni_lat)
                     border_rules_check.set_value(counter, "Uni Lng", uni_lng)
                     border_rules_check.set_value(counter, "CityState", city_state)
                     counter = counter + 1
                 else:
                     continue

border_rules_check.head()

border_rules_check.head()

#create additional dataframe by checking universities using university rules
uni_rules_check = pd.DataFrame(columns=['City', 'State', 'City Lat', 'City Lng',
                                      'University', 'Uni Lat', 'Uni Lng', 'Distance', 'Priority'])

counter = 0
for index, row in physical_locations.iterrows():
    print("Checking " + row["University"])
    uni_pull = row["University"]
    uni_lat = row["Uni_lat"]
    uni_lng = row["Uni_lng"]
    uni_name = row["University"]
    uni_loc = row["CityState"]
    for item in uni_rules[uni_pull]:
        if item is not "":
            #check against the city/state dataframe, create new dataframe of values, add values for city/state and lat/lng
            for index, row in full_cities.iterrows():
                if row["state"] == item or row["CityState"] == item:
                    city_name = row["city"]
                    state_name = row["state"]
                    city_lat = row["lat"]
                    city_lng = row["lng"]
                    city_state = row["CityState"]
                    
                    p1 = (uni_lat, uni_lng)
                    p2 = (city_lat, city_lng)
                    mi_distance = vincenty(p1, p2).miles

                    uni_rules_check.set_value(counter, "Distance", mi_distance)
                    uni_rules_check.set_value(counter, "City", city_name)
                    uni_rules_check.set_value(counter, "State", state_name)
                    uni_rules_check.set_value(counter, "City Lat", city_lat)
                    uni_rules_check.set_value(counter, "City Lng", city_lng)
                    uni_rules_check.set_value(counter, "University", uni_name)
                    uni_rules_check.set_value(counter, "Uni CityState", uni_loc)
                    uni_rules_check.set_value(counter, "Uni Lat", uni_lat)
                    uni_rules_check.set_value(counter, "Uni Lng", uni_lng)
                    uni_rules_check.set_value(counter, "CityState", city_state)
                    uni_rules_check.set_value(counter, "Priority", True)
                    counter = counter + 1

                else:
                    continue

uni_rules_check.head()

#university rules incorporated at front end, would then concatenate these frames and continue with rest of the script
frames = [border_rules_check, uni_rules_check]
additive_rules_df = pd.concat(frames, join='outer')

additive_rules_df.reset_index(inplace=True)

additive_rules_df.to_csv("full_additive_rules.csv")

additive_rules_df.head()

### Read the Negative Rules Files

#incorporate negative rules on full frame
#import university rules csv, clean up
negative_rules = {}
csvpath = 'Raw/negative-rules.csv'

with open(csvpath, newline= "") as csvfile:
    csvreader = csv.reader(csvfile, delimiter=',')
    next(csvreader)

    for row in csvreader:
        uni = row[0]
        negative_rules[uni] = row[1:]

for item in negative_rules:
    negative_rules[item] = [x for x in negative_rules[item] if x]

negative_rules

for index, row in additive_rules_df.iterrows():
    if row["University"] in negative_rules:
        for item in negative_rules[row["University"]]:
            if (row["State"] == item) or (row["CityState"] == item):
                additive_rules_df.drop(index, inplace=True)

test = additive_rules_df[additive_rules_df["University"] == "GTECH"]

test[test["State"] == "FL"]

additive_rules_df = additive_rules_df[['City', 'City Lat', 'City Lng', 'CityState', 'Distance',
       'Priority', 'State', 'Uni CityState', 'Uni Lat', 'Uni Lng',
       'University']]

additive_rules_df.to_csv("with_negative_rules.csv")

### Distance Check

additive_rules_df = additive_rules_df[additive_rules_df["Distance"] <= 200]

additive_rules_df.head()

### Export

#the following selection of CityStates is within range for a google distance check

final_rules = additive_rules_df

final_rules.to_csv("testing_distance.csv")

final_rules.to_sql("Rules_Added", conn, if_exists="replace", index=False)

# Check Tables in SQLite
cur = conn.cursor() 
res = cur.execute("SELECT name FROM sqlite_master WHERE type='table' ORDER BY name;")
for name in res:
    print(name[0])

final_rules[final_rules["CityState"] == "SHINGLEHOUSE, PA"]

### Export to Cloud SQL

engine = create_engine('mysql+mysqldb://trilogy:DataRocks@35.227.28.228/mapping_data?unix_socket=/cloudsql/sql-projects:us-east1:opportunity-db')
conn = engine.connect()

final_rules.to_sql("Rules_Added", conn, if_exists="replace", index=False)
