==> 01-Data_Prep/05-Impressions_Data.ipynb <==
## Impressions Data Retrieval
---
In this Jupyter notebook, we extract Impressions data on every city in our dataset.

# Dependencies
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import requests
import time 
from census import Census
from us import states
import sqlite3
from sqlalchemy import create_engine
import pymysql
pymysql.install_as_MySQLdb()

# Census API Key
c = Census("85ac64b6b5a9c0901b00329d1ef41f0c53ccfc98", year=2015)

### Load the City Census Data

conn = sqlite3.connect("Opportunity_Map.db")
city_census = pd.read_sql("select * from City_Census", conn)
city_census.head()

### Load the Impression Data

impressions_pd = pd.read_excel("Raw/Impressions_Data.xlsx")

impressions_pd.head()

### Clean the Impression Key

impressions_pd["CityState"] = impressions_pd["CityState"].str.upper()

impressions_pd["CityState"].head()

### Attempt Merge (for Testing)

census_impressions_merge = pd.merge(impressions_pd, city_census, on="CityState", how="inner")

impressions_pd.count()

city_census.count()

census_impressions_merge.count()

### Store the Impressions Data

impressions_pd.to_csv("Outputs/04-Impressions.csv")

impressions_pd.to_sql("Impressions", conn, if_exists="replace", index=False)

# Check Tables in SQLite
cur = conn.cursor() 
res = cur.execute("SELECT name FROM sqlite_master WHERE type='table' ORDER BY name;")
for name in res:
    print(name[0])

### Export to Cloud SQL

engine = create_engine('mysql+mysqldb://trilogy:DataRocks@35.227.28.228/mapping_data?unix_socket=/cloudsql/sql-projects:us-east1:opportunity-db')
conn = engine.connect()

impressions_pd.to_sql("Impressions", conn, if_exists="replace", index=False)
