==> 01-Data_Prep/06-Fully_Compiled_Data.ipynb <==
## Compile All Data Into A Single CSV

In this Jupyter Notebook, we compile results across all of our datasets in Part 1 - 5 into a single CSV. (clean up)

# Dependencies
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import requests
import time 
from census import Census
from us import states
import sqlite3
from sqlalchemy import create_engine
import pymysql
pymysql.install_as_MySQLdb()

# Census API Key
c = Census("85ac64b6b5a9c0901b00329d1ef41f0c53ccfc98", year=2015)

### Load All Databases

# Check Tables in SQLite
conn_lite = sqlite3.connect("Opportunity_Map.db")
cur = conn_lite.cursor() 
res = cur.execute("SELECT name FROM sqlite_master WHERE type='table' ORDER BY name;")
for name in res:
    print(name[0])

# Retrieve all data
city_census = pd.read_sql("select * from City_Census", conn_lite)
min_distances = pd.read_sql("select * from Minimum_Distances", conn_lite)
impressions = pd.read_sql("select * from Impressions", conn_lite)

min_distances.head()

city_census.head()

### Merge All Data

merged_data = pd.merge(city_census, min_distances, on=["CityState"], how="inner")

merged_data.columns

merged_data = pd.merge(merged_data, impressions, on="CityState", how="left")

merged_data.columns

merged_data.head()

merged_data.count()

merged_data.columns

mid_data = merged_data[['CityState', 'city', 'state', 'Population_x', 'White Population_x',
       'Black Population_x', 'Native American Population_x',
       'Asian Population_x', 'Hispanic Population_x', 'Education None_x',
       'Education High School_x', 'Education GED_x', 'Education Associates_x',
       'Education Bachelors_x', 'Education Masters_x',
       'Education Professional_x', 'Education Doctorate_x', 'Poverty_x',
       'Employment Labor Force_x', 'Employment Unemployed_x',
       'Employment Male Computer Engineering_x',
       'Employment Female Computer Engineering_x', 'Median Age_x',
       'Median Male Age_x', 'Median Female Age_x', 'Household Income_x',
       'Income Per Capita_x', 'Median Gross Rent_x', 'Median Home Value_x',
       'lat', 'lng', 'University', 'Uni CityState', 'Distance', 'Closest', 'City', 'City Lat', 'City Lng', 'Distance_Text',
       'State', 'Time', 'Time_Text', 'Uni Lat', 'Uni Lng', 'Impressions', 'CPC', 'CTR', 'Cost']]

mid_data[["lat", "City Lat"]]

final_data = mid_data[['CityState', 'Population_x', 'White Population_x',
       'Black Population_x', 'Native American Population_x',
       'Asian Population_x', 'Hispanic Population_x', 'Education None_x',
       'Education High School_x', 'Education GED_x', 'Education Associates_x',
       'Education Bachelors_x', 'Education Masters_x',
       'Education Professional_x', 'Education Doctorate_x', 'Poverty_x',
       'Employment Labor Force_x', 'Employment Unemployed_x',
       'Employment Male Computer Engineering_x',
       'Employment Female Computer Engineering_x', 'Median Age_x',
       'Median Male Age_x', 'Median Female Age_x', 'Household Income_x',
       'Income Per Capita_x', 'Median Gross Rent_x', 'Median Home Value_x','University', 'Uni CityState', 'Distance', 
        'Closest', 'City', 'City Lat', 'City Lng', 'Distance_Text',
       'State', 'Time', 'Time_Text', 'Uni Lat', 'Uni Lng', 'Impressions', 'CPC', 'CTR', 'Cost']]

final_data = final_data.rename(columns = {'Population_x':"Population", 'White Population_x':"White Population",
       'Black Population_x':"Black Population", 'Native American Population_x':"Native American Population",
       'Asian Population_x':"Asian Population", 'Hispanic Population_x':"Hispanic Population", 'Education None_x':"Education None",
       'Education High School_x':"Education High School", 'Education GED_x':"Education GED", 'Education Associates_x':"Education Associates",
       'Education Bachelors_x':"Education Bachelors", 'Education Masters_x':"Education Masters",
       'Education Professional_x':"Education Professional", 'Education Doctorate_x':"Education Doctorate", 'Poverty_x':"Poverty",
       'Employment Labor Force_x':"Employment Labor Force", 'Employment Unemployed_x':"Employment Unemployed",
       'Employment Male Computer Engineering_x':"Employment Male Computer Engineering",
       'Employment Female Computer Engineering_x':"Employment Female Computer Engineering", 'Median Age_x':"Median Age",
       'Median Male Age_x':"Median Male Age", 'Median Female Age_x':"Median Female Age", 'Household Income_x':"Household Income",
       'Income Per Capita_x':"Income Per Capita", 'Median Gross Rent_x':"Median Gross Rent", 'Median Home Value_x':"Median Home Value"})

final_data.columns

final_data = final_data[['CityState', 'University', 'Uni CityState', 'Distance', 'Closest', 'City', 'City Lat', 'City Lng',
       'Distance_Text', 'State', 'Time', 'Time_Text', 'Uni Lat', 'Uni Lng','Population', 'White Population', 'Black Population',
       'Native American Population', 'Asian Population', 'Hispanic Population',
       'Education None', 'Education High School', 'Education GED',
       'Education Associates', 'Education Bachelors', 'Education Masters',
       'Education Professional', 'Education Doctorate', 'Poverty',
       'Employment Labor Force', 'Employment Unemployed',
       'Employment Male Computer Engineering',
       'Employment Female Computer Engineering', 'Median Age',
       'Median Male Age', 'Median Female Age', 'Household Income',
       'Income Per Capita', 'Median Gross Rent', 'Median Home Value',
       'Impressions', 'CPC', 'CTR', 'Cost']]

final_data

### Export Data to Local DB

final_data.to_csv("Outputs/06-Final_Data.csv")

final_data.to_sql("Full_Data", conn_lite, if_exists="replace", index=False)

# Check Tables in SQLite
cur = conn_lite.cursor() 
res = cur.execute("SELECT name FROM sqlite_master WHERE type='table' ORDER BY name;")
for name in res:
    print(name[0])

### Export to Cloud SQL

# Create Engine and Pass in MySQL Connection
engine = create_engine('mysql+mysqldb://trilogy:test@35.227.28.228/mapping_data?unix_socket=/cloudsql/sql-projects:us-east1:opportunity-db')
conn = engine.connect()

final_data.to_sql("Full_Data", conn, if_exists="replace", index=False)

data = pd.read_sql("SELECT * FROM Full_Data", conn)

data.head()

#gtech_df = data[data["University"] == "GTECH"]

#gtech_df.to_csv("gtech_area.csv")
